requirements.txt
torch
numpy
soundfile
loadenv
torchaudio
numpy
scipy
scikit-learn
pandas
seaborn
matplotlib
tensorboard
transformers
google
pathlib
librosa
noisereduce
dotenv
pyclarity
openai-whisper
inflect
jiwer
hydra-core
pystoi